[up:objective-functions-losses] rename initializers -> init, objectives -> objective
