+++ nujo/init/random.py
